+++ SimulacrumBossStageMod/CHANGELOG.md
@@ -1,3 +1,8 @@
+# 1.3.2
+
+- Make it possible to have Artifacts enabled from the start of the run
+- Bosses do not take void fog damage
+
 # 1.3.1
 
 - Tweak on the boss wave values